ai-service/requirements.txt
fastapi
uvicorn
openai-whisper
transformers
torch
python-multipart
pymongo
python-dotenv
dateparser
